feat(api): add auth to api end points

+++ graphql/schema/mutations/delete-link.ts
@@ -5,6 +5,10 @@ export const deleteLink = mutationField('deleteLink',{
   type: Link,
   args: { id: nonNull(intArg())},
   resolve(_parent, { id }, ctx) {
+    const { userId } = ctx
+    if (!userId) {
+      throw new Error('must be logged in to delete posts')
+    }
     return ctx.prisma.link.delete({
       where: { id },
     })
